Trying to detect if server is actually there...

+++ bb.js
@@ -23,8 +23,12 @@ let blnNextHTTPpost = true, objStats = {}, options
 objStats[cmn.defs.JSON_BEAGLEBONE] = {}
 cmn.registration(objStats, cmn.defs.JSON_BEAGLEBONE, 7, 1)
 //Remove any existing pipe files
-cmn.fs.unlinkSync(cmn.defs.PIPE_ASAFE_TO_BB)
-cmn.fs.unlinkSync(cmn.defs.PIPE_BB_TO_ASAFE)
+if ( cmn.fs.existsSync(cmn.defs.PIPE_ASAFE_TO_BB) ) {
+    cmn.fs.unlinkSync(cmn.defs.PIPE_ASAFE_TO_BB)
+}
+if ( cmn.fs.existsSync(cmn.defs.PIPE_BB_TO_ASAFE) ) {
+    cmn.fs.unlinkSync(cmn.defs.PIPE_BB_TO_ASAFE)
+}
 //Report profile folder
 cmn.displayDataWithPrompt(5, 6, process.env.HOME, "Profile: ")
 cmn.displayDataWithPrompt(6, 1, strCWD, "Current path: ")
